[median of two sorted arrays] Bruth force solution.

+++ median-of-two-sorted-arrays/main_test.go
@@ -11,8 +11,8 @@ func TestMedianBruthForce(t *testing.T) {
   input1 := Input{arr1: []int{1,3}, arr2: []int{2}}
   input2 := Input{arr1: []int{1,2}, arr2: []int{3,4}}
 
-  expected1 := float32(2)
-  expected2 := float32(2.5)
+  expected1 := float64(2)
+  expected2 := float64(2.5)
 
   result1 := BruthForceSolution(input1.arr1, input1.arr2)
   result2 := BruthForceSolution(input2.arr1, input2.arr2)
